refactor: move PopularitySorting and throwIfEmpty to their associated packages
