+++ domain/src/main/java/com/example/domain/common/ContentFilteringExtensions/ThrowIfEmpty.kt
@@ -1,4 +1,4 @@
-package com.example.domain.common
+package com.example.domain.common.ContentFilteringExtensions
 
 fun <T> List<T>.throwIfEmpty(exception: () -> Exception): List<T> {
     return this.ifEmpty { throw exception() }
